[IMP] runbot: grant rights for the build logs
When an Odoo instance is launched buy a build, the --log-db parameter is
used to write its logs into the runbot ir_logging table.
Also, a trigger on the postgresql server is used to add the build id on
the ir_logging.

The the postgresql trigger also modifies the log_counter and
trigerred_result columns on the runbot_build table.

Because of that, the user used in the --log-db parameter must have the
rights to do those operations. Before this commit, it was up to the
runbot administrator to properly setup those rights.

With this commit, the rights are automatically granted when the logdb
uri is set or changed in the runbot settings.

The migration script for 4.5 is also adapted.