[IMP] runbot: create a temporary symlink in Docker container
When odoo saves files in '/tmp' during a build, those files are lost
when the container is stopped. Typically screenshots during the HttpCase
tests.

With this commit, a directory 'tests' is created in the build dir and a
symlink is also created in /tmp, pointing to it.

With this mechanism, the files saved in '/tmp/odoo_tests' automatically
goes in buildir and survives the container removal.