[IMP] runbot: add dbfread to the Docker image
The new feature in odoo/enterprise#4892 needs the dbfread module.
As this lib is not required for other Odoo modules, it cannot be added
to the requirements.txt file.

In order to run the tests and use this new feature on the runbot, this
commit installs the dbfread lib in the Docker image.