[FIX][IMP] runbot: use environament variable for fqdn
In some conditions, the getfqdn method does not return the expected
fqdn. e.g. when the local fqdn gives something else than the resolved
fqdn.

Also, in a near future, the runbot should not depend on the fqdn
anymore, in order to manually assign host identification.

With this commit, fqdn is still used unless a RUNBOTNAME environment
variable exists.