[FIX] forwardbot: selection of ancestor in FP tail ping
In selecting the parent commits to list on the last PR, we would miss
the *first* forward-port of the sequence. Not sure why we added a
detrimental check on source_id there.

Also add a missing space between "chain" and "containing" in the case
where there's at least one forward-port PR other than the final one.

Fixes #212